backend-python/requirements.txt
fastapi
uvicorn
asyncpg
redis
python-jose[cryptography]
passlib[bcrypt]
python-dotenv
python-socketio
pydantic
pydantic-settings
httpx
geopy
loguru
